Create a directory for other pages with recipes
Create a directory which will contain the other pages of the website
where the recipes of the dishes will be described.
Add a html file for a pizza recipe containing the usual boilerplate and
a header with the name of the dish.
In index.html add the relative link to the pizza.html.